feat: move helpers up a level for reuse purposes
