+++ pkg/plugin/helpers.go
@@ -1,4 +1,4 @@
-package postgresql
+package plugin
 
 import (
 	v1beta12 "halkyon.io/api/capability/v1beta1"
@@ -7,12 +7,12 @@ import (
 	"strings"
 )
 
-func ownerAsCapability(res framework.DependentResource) *v1beta12.Capability {
+func OwnerAsCapability(res framework.DependentResource) *v1beta12.Capability {
 	return res.Owner().(*v1beta12.Capability)
 }
 
 // Convert Array of parameters to a Map
-func parametersAsMap(parameters []v1beta1.NameValuePair) map[string]string {
+func ParametersAsMap(parameters []v1beta1.NameValuePair) map[string]string {
 	result := make(map[string]string)
 	for _, parameter := range parameters {
 		result[parameter.Name] = parameter.Value
@@ -54,8 +54,13 @@ func SetDefaultDatabasePort(paramPort string) string {
 }
 
 //getAppLabels returns an string map with the labels which wil be associated to the kubernetes/ocp resource which will be created and managed by this operator
-func getAppLabels(name string) map[string]string {
+func GetAppLabels(name string) map[string]string {
 	return map[string]string{
 		"app": name,
 	}
+}
+
+func ReplicaNumber(num int) *int32 {
+	q := int32(num)
+	return &q
 }
